refactor: depreacte websockets
