--- packages/maumau-server/src/server/websocket.d.ts
@@ -1,5 +1,0 @@
-import WS from 'ws';
-
-interface CustomSocket extends WS {
-  isAlive: boolean;
-}
